+++ lib/ansible/ruby/models/playbook_spec.rb
@@ -70,6 +70,33 @@ describe Ansible::Ruby::Models::Playbook do
     end
   end
 
+  context 'includes' do
+    let(:instance) do
+      Ansible::Ruby::Models::Playbook.new plays: [play1],
+                                          inclusions: [Ansible::Ruby::Models::Inclusion.new(file: 'something.yml')]
+    end
+
+    it do
+      is_expected.to eq [
+        {
+          hosts: 'host1:host2',
+          name: 'play 1',
+          tasks: [
+            {
+              name: 'do stuff on EC2',
+              ec2: {
+                foo: 123
+              }
+            }
+          ]
+        },
+        {
+          include: 'something.yml'
+        }
+      ]
+    end
+  end
+
   context 'no plays' do
     let(:instance) do
       Ansible::Ruby::Models::Playbook.new