+++ lib/ansible/ruby/modules/custom/commands/script.rb
@@ -1,7 +1,7 @@
 # VALIDATED_CHECKSUM: RhHDiVtIFNT3I5ESwjmmuJBl8pv6qi2oIyGxPsjxmpE=
 # See LICENSE.txt for license
 
-require 'ansible/ruby/modules/custom/free_form'
+require 'ansible/ruby/modules/free_form'
 require 'ansible/ruby/modules/generated/core/commands/script'
 
 module Ansible