+++ spec/rake/nested_tasks/playbook1_test.rb
@@ -5,3 +5,5 @@ play 'the play name' do
 
   user 'centos'
 end
+
+ansible_include 'other_playbook.yml'