+++ spec/rake/no_nested_tasks/playbook1_test.rb
@@ -1,6 +1,7 @@
 play 'the play name' do
   hosts %w(host1 host2)
 
+  ansible_include 'included_file.yml'
   task 'Copy something over' do
     copy do
       src '/file1.conf'